似乎能复用查询，但是感觉问题是不是有点大？为了查一个数字去返回整张表，性能应该不是很好，但是如果不这样写感觉好不优雅，我之前好不容易统一的格式哇...

+++ Redrock-student-system/go/dto/homework_res.go
@@ -24,7 +24,7 @@ type HomeworkItem struct {
 	Creator         CreatorInfo `json:"creator"`
 	Deadline        time.Time   `json:"deadline"`
 	AllowLate       bool        `json:"allow_late"`
-	SubmissionCount int         `json:"submission_count"`
+	SubmissionCount uint64      `json:"submission_count"`
 }
 type FindHomeworkRes struct {
 	List     []HomeworkItem `json:"list"`
